requirements.txt
Pillow==9.2.0
streamlit>=1.10.0
streamlit-drawable-canvas==0.9.1
torch==1.12.0
torchvision==0.13.0
gdown